Rework Products/Stock split, decimal box ordering, balance timing, stock cycle time
- Products page no longer has any price field — it's purely for creating a product for future use (cost price stays, optional). All pricing now happens in Stock.
- Stock page redesigned: a "Needs Pricing" section lists everything bought that hasn't gone live, lets the admin enter quantity in pallets or boxes (auto-converted to boxes) plus price per line, mark each "Done", then a single "Release Prices to Customers" button publishes them all at once
- Customers no longer see exact pallet/box counts anywhere — just Good Stock/Low Stock/Out of Stock, plus a small "X boxes per pallet" note when they view a product
- Order quantity is now entered in boxes and accepts decimals (e.g. 2.5 boxes) instead of whole units
- Daily stock cycle time changed from 06:00 to 10:00 UK time everywhere (copy + logic)
- Buying Desk: "End Daily Buying" can now be run again the same day after more buying is confirmed — relabelled "Send Extra Buying to Stock" for that case, same trading date, just a later session
- Fixed balance not updating: pay-before-order (or over-limit) sales now update the customer's balance the moment an admin confirms the order; credit-term sales are invoiced and added to balance only when Day End is run for that trading date. Paying an invoice (manual mark-paid or approved payment proof) now correctly brings the balance back down. All verified live end-to-end except Day End itself, which is destructive/irreversible on production data so was verified by code review instead.
- Needs one migration: stock_items.available_quantity changed from integer to numeric(10,2) to allow decimal box quantities

+++ src/lib/schema.sql
@@ -132,13 +132,17 @@ create table if not exists products (
 create table if not exists stock_items (
   id                 text primary key default gen_random_uuid()::text,
   product_id         text references products(id) on delete cascade,
-  available_quantity integer default 0,
+  -- Always stored as boxes — decimal so a customer can order e.g. 2.5 boxes.
+  available_quantity numeric(10,2) default 0,
   price              numeric(10,2) default 0,
   last_updated       text,
   status             text default 'available',
   packaging          text,
   created_at         timestamptz default now()
 );
+
+-- If stock_items already existed with an integer available_quantity, run:
+-- alter table stock_items alter column available_quantity type numeric(10,2);
 
 -- ORDERS
 create table if not exists orders (
